Refac: move code and apply ESLint (#7)
